Updated
Changed the output of all the tasks to the new pattern with output.
Changed to specific IDs where appropriate and also full JSON payload
where appropriate. Added end to end template example with provisioning
Skytap VM and mapping that to an XLD environment, deploying, and then
tearing everything down and cleaning up.